==> code/2 - data_wrangling_ensanut_2018.Rmd <==
---
title: "2 - data_wrangling_ensanut_2018"
author: "Luis E Segura"
date: "2023-11-07"
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)

options(tinytex.verbose = TRUE, scipen = 999)

mypckgs <- c("tidyverse", "here")

for (package in mypckgs) {
  library(package, character.only = T)
}

load(here("data/", "ensanut18_all.RData"))
```

Limpiamos y recodificamos variables

```{r}
ensanut18_4analysis <- ensanut18_all |>
  labelled::remove_labels()  |>
  mutate(ses = factor(case_when(nsef == 1 ~ 3, 
                                nsef == 2 ~ 2, 
                                nsef == 3 ~ 1), 
                      labels = c("High", "Middle", "Low")), 
         age_cat = factor(ifelse(edad < 40, 1, 
                                 ifelse(edad > 39 & edad < 60, 2, 3)), 
                          labels = c("20 - 39 yo", "40 - 59 yo", "60+ yo")), 
         sexo = factor(sexo, labels = c("Men", "Women")), 
         urbanicity = factor(dominio, labels = c("Urban", "Rural")),  
         region = factor(region, labels = c("North", "Center", "Mexico City", "South")),
         region = fct_relevel(region, "Center"),
         l_indigena = factor(ifelse(l_indigena == 2, 0, l_indigena), labels = c("No", "Yes")), 
         diabetes = factor(ifelse(diabetes > 1, 0, diabetes), labels = c("No", "Yes")),
         diab_check = factor(case_when(diab_check %in% c("01", "02") ~ "IMSS", 
                          diab_check %in% c("03", "04", "11") ~ "SS",
                          diab_check %in% c("05") ~ "DIF y CR",
                          diab_check %in% c("06", "07") ~ "ISSSTE",
                          diab_check == "08" ~ "Marina, Defensa, PEMEX",
                          diab_check == "09" ~ "Privado",
                          diab_check == "10" ~ "Hospital Civil",
                          diab_check == "12" ~ "Consultorios de Farmacias",
                          diab_check %in% c("13", "14") ~ "Otras")),
         diab_rev = factor(ifelse(diab_rev == 2, 0, 
                           ifelse(diab_rev == 9, NA, diab_rev)), labels = c("No", "Yes")), 
         diab_dv = factor(ifelse(diab_dv == 2, 0, diab_dv), labels = c("No", "Yes")),
         diab_perd_vsn = factor(ifelse(diab_perd_vsn == 2, 0, diab_perd_vsn), labels = c("No", "Yes")),
         hipertension = factor(ifelse(hipertension == 2, 0, hipertension), labels = c("No", "Yes")),
         diab_hbp = factor(case_when(diabetes == "No" & hipertension == "No" ~ 1, 
                              diabetes == "Yes" & hipertension == "No" ~ 2, 
                              diabetes == "No" & hipertension == "Yes" ~ 3, 
                              diabetes == "Yes" & hipertension == "Yes" ~ 4), 
                           labels = c("No diabetes or HTN", "Diabetes no HTN", "HTN no diabetes", "Both diabetes & HTN")),
         hta_retina = factor(ifelse(hta_retina == 2, 0, hta_retina), labels = c("No", "Yes")),
         usa_lentes = factor(ifelse(usa_lentes == 2, 0, usa_lentes), labels = c("No", "Yes")),
         disc_visual = factor(disc_visual, labels = c("No difficulty", "Some difficulty", 
                                                      "A lot of difficulty", "Cannot see at all")),
         disc_visual_severa = factor(ifelse(disc_visual %in% c("No difficulty", "Some difficulty"), 0, 1), labels = c("No", "Yes")),
         disc_visual_severa_num = ifelse(disc_visual %in% c("No difficulty", "Some difficulty"), 0, 1),
         p2_1_1 = p2_1_1 - 1, 
         p2_1_2 = p2_1_2 - 1, 
         p2_1_3 = p2_1_3 - 1, 
         p2_1_4 = p2_1_4 - 1, 
         p2_1_5 = p2_1_5 - 1, 
         p2_1_6_r = case_when(p2_1_6 == 4 ~ 3, 
                              p2_1_6 == 3 ~ 2, 
                              p2_1_6 == 2 ~ 1, 
                              p2_1_6 == 1 ~ 0), 
         p2_1_7 = p2_1_7 - 1, 
         cesd_score = rowSums(pick(p2_1_1, p2_1_2, p2_1_3, p2_1_4, p2_1_5, p2_1_6_r, p2_1_7)), 
         cesd = factor(ifelse((cesd_score > 8 & edad < 60) | (cesd_score > 4 & edad > 59), 1, 0), 
                       labels = c("No depressive symptoms", "Depressive symptoms")),
         obesidad = factor(ifelse(obesidad == 2, 0, obesidad), labels = c("No", "Yes")),
         hrt_attck = ifelse(p5_2_1 == 2, 0, p5_2_1),
         angina = ifelse(p5_2_2 == 2, 0 , p5_2_2),
         hrt_fail = ifelse(p5_2_3 == 2, 0, p5_2_3),
         cardio_dis = factor(ifelse(hrt_attck ==  1 | angina == 1 | hrt_fail == 1, 1, 0), labels = c("No CVD", "Yes CVD")),
         kidney_dis = factor(ifelse(e_renal_crn == 2, 0, e_renal_crn), labels = c("No", "Yes")),
         dyslipidemia = factor(ifelse(p6_4 == 1 | p6_6 == 1, 1, 0), labels = c("No", "Yes")),
         smk_ever = factor(ifelse(p13_1 == 2, 0, 
                           ifelse(p13_1 == 9, NA, p13_1)), labels = c("No", "Yes")),
         smk_curr = factor(ifelse(p13_2 == 3, 0, 
                                  ifelse(p13_2 %in% c(1, 2), 1, NA)), labels = c("No", "Yes")), 
         smk_former = factor(ifelse(p13_4 == 3, 0, 
                                    ifelse(p13_4 %in% c(1, 2), 1, NA)), labels = c("No", "Yes")),
         smk_cat = factor(case_when(p13_1 == 1 & p13_2 == 1 ~ 4, 
                             p13_1 == 1 & p13_2 == 2 ~ 3, 
                             p13_1 == 1 & p13_2 == 3 ~ 2 , 
                             p13_1 == 2 ~ 1), 
                          labels = c("Never smokers", "Former smokers", "Current somedays", "Current daily")), 
         p13_14 = as.integer(p13_14), 
         binge_drnk = factor(case_when(p13_14 == 88 ~ NA,
                                p13_14 > 0 & p13_14 < 81 ~ 1, 
                                p13_14 %in% c(0, 98) ~ 0), labels = c("No", "Yes")), 
         heavy_drnk = case_when(p13_14 > 88 ~ NA,
                                p13_14 > 4 & p13_14 < 81 ~ 1, 
                                p13_14 < 5 ~ 0, 
                                p13_14 == 98 ~ 2), 
         alcohol = factor(case_when(p13_11 == 3 ~ 0,
                                    p13_11 == 2 ~ 1, 
                                    p13_11 == 1 ~ 2), labels = c("Never", "No", "Yes")), 
         work_pwk = factor(case_when(trabajo == 2 ~ 1, 
                                     trabajo == 1 ~ 0), labels = c("Yes", "No")), 
         insurance = factor(ifelse(p3_10_99 == 1, NA, 
                                   ifelse(p3_10_11 == 1, 0, 1)), labels = c("No insurance", "Insurance")), 
         work_type = factor(tipo_trabajo, labels = c("empleada(o)", "obrera(o)", "jornalera(o) o peona(ón)", 
                                              "trabajadora(or) por su cuenta", "patrona(ón) o empleadora(or)", 
                                              "trabajadora(or) sin pago", "otro")), 
         nivel = as.integer(nivel),
         education = factor(case_when(nivel < 3 ~ 4, 
                                      nivel == 3 ~ 3, 
                                      nivel == 4 ~ 2, 
                                      nivel > 4 ~ 1), 
                            labels = c("More than high school education", 
                                       "High school education","Secondary school", 
                                       "Elementary education or lower")), 
         accidente = factor(ifelse(accidente == 2, 0, accidente), labels = c("No", "Yes")), 
         dis_accidente = factor(case_when(disc_visual_accdnt == 7 ~ 0, 
                                   disc_visual_accdnt %in% c(1, 2) ~ 1, 
                                   disc_visual_accdnt %in% c(3, 4) ~ 2, 
                                   disc_visual_accdnt == 5 ~ 3, 
                                   disc_visual_accdnt == 6 ~ 4, 
                                   disc_visual_accdnt == 9 ~ NA), 
                                labels = c("None (REF)", "Motor disability", "Speaking/hearing disability", 
                                           "Visual disability/Blindness", "Other physical/mental disability")), 
         accidente_lim = factor(case_when(accidente == "No" & is.na(dis_accidente) ~ 1, 
                                   accidente == "Yes" & dis_accidente == "None (REF)" ~ 2, 
                                   accidente == "Yes" & dis_accidente == "Motor disability" ~ 3, 
                                   accidente == "Yes" & dis_accidente == "Speaking/hearing disability" ~ 4, 
                                   accidente == "Yes" & dis_accidente == "Visual disability/Blindness" ~ 5,
                                   accidente == "Yes" & dis_accidente == "Other physical/mental disability" ~ 6), 
                                labels = c("No accident", "No disability due to accident", "Motor disability", 
                                           "Hearing/speaking disability", 
                                           "Visual disabsility/blindness", "Other physical/mental disability"))) 

ensanut18_4analysis <- ensanut18_4analysis |>
  select(upm:ent, region, est_dis:f_20mas, age_cat, sexo, urbanicity, education, ses, insurance, l_indigena, work_pwk, work_type, diabetes, 
         diab_check, diab_rev, diab_dv, diab_perd_vsn, hipertension, hta_retina, usa_lentes, disc_visual, disc_visual_severa, disc_visual_severa_num, 
         cesd_score, cesd, obesidad, cardio_dis, kidney_dis, dyslipidemia, smk_cat, alcohol, binge_drnk, accidente, dis_accidente, accidente_lim, diab_hbp)


save(ensanut18_4analysis, file = here("data/", "ensanut18_4analysis.RData"))


```
