==> code/4 - Asociaciones predictores de SVD.Rmd <==
---
title: "4 - Asociaciones predictores de SVD Supplemental Material"
author: "Luis E Segura"
date: "2023-12-13"
output: html_document
---

```{r, include=FALSE}
knitr::opts_chunk$set(echo = FALSE, warning = FALSE, message = FALSE)

options(tinytex.verbose = TRUE, scipen = 999)

mypckgs <- c("tidyverse", "here", "survey", "srvyr", "doParallel", "foreach")

for (package in mypckgs) {
  library(package, character.only = T)
}

## setup parallel processing
no_cores <- detectCores() - 1  
registerDoParallel(cores = no_cores)
options(cores = no_cores)
getDoParWorkers()                   
getDoParName()

```

```{r}
### leemos la base de datos
load(here("data/", "ensanut18_4analysis.RData"))
```

```{r}
### Creamos el objeto del diseño muestral para calcular los estimadores ponderados (p.ej., prevalencias)
ensanut_diseno <- ensanut18_4analysis |>
  as_survey_design(strata = est_dis, 
                   ids = upm_dis, 
                   weights = f_20mas, 
                   nest = T)
```

```{r}
reg_fun <- function(x, y){
  
  reference <- ensanut18_4analysis |>
    pull(!!sym(x)) |> 
    levels()
  
  reference <- reference[1]
  
  my_formula <- as.formula(str_glue("{y} ~ {x}"))
  
  svyglm(my_formula, design = ensanut_diseno, family = quasipoisson) |>
    broom::tidy(exp = T, conf.int = T) |>
    select(term, PR = estimate, conf.low, conf.high, p.value) |>
    mutate(var = x, 
           PR = ifelse(term == "(Intercept)", 1, PR), 
           conf.low = ifelse(term == "(Intercept)", NA, conf.low), 
           conf.high = ifelse(term == "(Intercept)", NA, conf.high), 
           p.value = ifelse(term == "(Intercept)", NA, p.value), 
           term = ifelse(term == "(Intercept)", str_glue("{reference} (REF)"), 
                         str_sub(term, start = str_length(x) + 1, end = str_length(term)))) |>
    relocate(var, .before = "term")
           
}
```


```{r}
var_char <- names(ensanut18_4analysis)[c(6, 8, 13:19, 44)]


tab_estimates <- foreach(i = var_char, .packages = c("tidyverse", "survey", "srvyr")) %dopar% {
  reg_fun(i, "disc_visual_severa_num")

}

```

```{r}
tab_est_coefs <- tab_estimates |>
  bind_rows() 

# write_csv(tab_est_coefs, here("results/", "socio_dem_coefs.csv"))

### Figure 1. Sociodemographic predictor of Severe Visual Impairment among the Mexican population.
tab_est_coefs |>
  mutate(term = case_when(var == "l_indigena" & term == "No (REF)" ~ "Doesn't speak indigenous language (REF)", 
                          var == "l_indigena" & term == "Yes" ~ "Speaks indigenous language", 
                          var == "insurance" & term == "No insurance (REF)" ~ "Uninsured (REF)", 
                          var == "insurance" & term == "Insurance" ~ "Insured",
                          var == "work_pwk" & term == "Yes (REF)" ~ "Past-week employed (REF)", 
                          var == "work_pwk" & term == "No" ~ "Past-week unemployed",
                          var == "diab_hbp" & term == "No diabetes or HTN (REF)" ~ "Neither diabetes nor hypertension (REF)", 
                          var == "diab_hbp" & term == "Diabetes no HTN" ~ "Diabetes / no hypertension",
                          var == "diab_hbp" & term == "HTN no diabetes" ~ "Hypertension / no diabetes",
                          var == "diab_hbp" & term == "Both diabetes & HTN" ~ "Diabetes & hypertension", 
                          T ~ term), 
         term =  factor(term),  
         orden = seq(1, length(var), 1),
         color = factor(c(rep(1, 2), rep(2, 4), rep(3, 3), rep(4, 2), rep(5, 4), 
                   rep(6, 3), rep(7, 2), rep(8, 2), rep(9, 2), rep(10, 4)), 
                   labels = c("Sex", "Region", "Age group", "Urbanicity", "Education", 
                              "SES", "Insurance", "Indigenous", "PW employment", 
                              "Diabetes and HTN"))) |>
  ggplot(aes(y = reorder(ordered(interaction(term, color, sep = "!")), desc(orden)), 
             x = PR, xmin = conf.low, xmax = conf.high, color = color)) +
  geom_point(size = 5, shape = "diamond") +
  geom_errorbarh(height = 0.3) + 
  scale_x_continuous(limits = c(0.2, 9.5), breaks = c(0.7, seq(1.0, 9.5, 1.0)), name = "Prevalence Ratio (95% CI)")  +
  scale_y_discrete(guide = ggh4x::guide_axis_nested(delim = "!")) +
  geom_vline(xintercept = 1, color = "black", linetype = "solid", alpha = 0.5) +
  ggthemes::theme_calc() +
  theme(legend.position = "none", 
        legend.title = element_blank(), 
        plot.title = element_text(size = 14), 
        text = element_text(size = 12)) +
  labs(y = "", 
       title = "")


# ggsave(here("results/", "forestplot_crude.pdf"), width = 12, height = 6, device = "pdf", units = "in", dpi = 1800)
# ggsave(here("results/", "forestplot_crude.jpeg"), width = 12, height = 8, units = "in", dpi = 1800)

```
